fixing linter errors on accounts show
