+++ webapp/src/routes/App/Accounts/index.ts
@@ -1,7 +1,9 @@
 import Index, { loader as indexLoader } from "./Index";
 import Layout from "./layout";
 import { loader as mainLoader } from "./loader";
-import Show, { action as showAction, loader as showLoader } from "./Show";
+import Show from "./show";
+import { action as showAction } from "./show/action";
+import { loader as showLoader } from "./show/loader";
 
 export default {
     loaders: {
